Move `BaseListItem` to `Icinga\Module\Icingadb\Common`
