+++ library/Icingadb/Widget/ItemList/BaseCommentListItem.php
@@ -12,7 +12,7 @@ use Icinga\Module\Icingadb\Common\NoSubjectLink;
 use Icinga\Module\Icingadb\Common\ObjectLinkDisabled;
 use Icinga\Module\Icingadb\Common\ServiceLink;
 use Icinga\Module\Icingadb\Model\Comment;
-use Icinga\Module\Icingadb\Widget\BaseListItem;
+use Icinga\Module\Icingadb\Common\BaseListItem;
 use Icinga\Module\Icingadb\Widget\TimeAgo;
 use ipl\Html\Attributes;
 use ipl\Html\BaseHtmlElement;
