+++ library/Icingadb/Widget/ItemList/BaseDowntimeListItem.php
@@ -13,7 +13,7 @@ use Icinga\Module\Icingadb\Common\NoSubjectLink;
 use Icinga\Module\Icingadb\Common\ObjectLinkDisabled;
 use Icinga\Module\Icingadb\Common\ServiceLink;
 use Icinga\Module\Icingadb\Model\Downtime;
-use Icinga\Module\Icingadb\Widget\BaseListItem;
+use Icinga\Module\Icingadb\Common\BaseListItem;
 use Icinga\Module\Icingadb\Widget\DowntimeList;
 use ipl\Html\Attributes;
 use ipl\Html\BaseHtmlElement;
